feat: Implement application routing
Add react-router-dom to handle navigation.

- Create routes for Main, Login, Favorites, and Offer pages.
- Add a 404 Not Found page for unknown paths.
- Implement a PrivateRoute component to protect the Favorites page.
