--- src/App.tsx
@@ -1,13 +1,0 @@
-import MainPage from './pages/main-page';
-
-type AppProps = {
-  offersCount: number;
-}
-
-function App({ offersCount }: AppProps): JSX.Element {
-  return (
-    <MainPage offersCount={offersCount} />
-  );
-}
-
-export default App;
