+++ src/app/app-router.tsx
@@ -1,0 +1,43 @@
+import { Route, Routes } from 'react-router-dom';
+import MainPage from '../pages/main-page/main-page';
+import LoginPage from '../pages/login-page/login-page';
+import FavouritesPage from '../pages/favourites-page/favourites-page';
+import OfferPage from '../pages/offer-page/offer-page';
+import NotFoundPage from '../pages/not-found-page/not-found-page';
+import { AppRoute } from './routes';
+import { PrivateRoute } from '../components/private-route/private-route';
+
+type AppRouterProps = {
+  offersCount: number;
+}
+
+export function AppRouter({ offersCount }: AppRouterProps): JSX.Element {
+  return (
+    <Routes>
+      <Route
+        path={AppRoute.Main}
+        element={<MainPage offersCount={offersCount} />}
+      />
+      <Route
+        path={AppRoute.Login}
+        element={<LoginPage />}
+      />
+      <Route
+        path={AppRoute.Favorites}
+        element={
+          <PrivateRoute isAuthorized={false}>
+            <FavouritesPage />
+          </PrivateRoute>
+        }
+      />
+      <Route
+        path={AppRoute.Offer}
+        element={<OfferPage />}
+      />
+      <Route
+        path={AppRoute.NotFound}
+        element={<NotFoundPage />}
+      />
+    </Routes>
+  );
+}
