+++ src/app/routes.ts
@@ -1,0 +1,7 @@
+export const enum AppRoute {
+  Main = '/',
+  Login = '/login',
+  Favorites = '/favorites',
+  Offer = '/offer/:id',
+  NotFound = '*',
+}
